refactor: component moved to its parent section since its very specific for it.
