+++ packages/src/sections/skills/carousel/hooks/useCarousel.ts
