+++ packages/src/sections/skills/carousel/utils/getMockCarousel.ts
